Feature - Implement eager loading relations.
Note: Still experimental, but seems to be pretty usable.

I first started down the route of creating a JOIN statement to load
all associations in one go. This proved difficult to extract the
proper rows back out and associate them within the proper object.

This will make one query per included relation in a query. The good
news is that queries are performed in parallel where possible.